[KWR001][D] code copy algorithm + UI

+++ nts.uk/uk.at/at.ctx/function/nts.uk.ctx.at.function.infra/src/main/java/nts/uk/ctx/at/function/infra/repository/dailyworkschedule/JpaOutputItemDailyWorkScheduleGetMemento.java
@@ -51,7 +51,7 @@ public class JpaOutputItemDailyWorkScheduleGetMemento implements OutputItemDaily
 	 */
 	@Override
 	public OutputItemSettingCode getItemCode() {
-		return new OutputItemSettingCode((int) this.kfnmtItemWorkSchedule.getId().getItemCode());
+		return new OutputItemSettingCode(this.kfnmtItemWorkSchedule.getId().getItemCode());
 	}
 
 	/* (non-Javadoc)
